feat: challenge api
Former-commit-id: 0611d4d31f2b6111d14ca4e9dc89edd4a8778dd9

+++ ondo/src/main/java/com/nextlevel/ondo/repository/ChallengeRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ChallengeRepository extends JpaRepository<Challenge, Integer> {
+public interface ChallengeRepository extends JpaRepository<Challenge, Long> {
     List<Challenge> findAllByCategory(Category category);
     Challenge findByChallengeId(long challengeId);
     List<Challenge> findByCategory(Category category);
